refactor(ai-engine): resolve #1730 — split online_research_agent + consolidate modrinth/curseforge parsers
