+++ ai-engine/agents/online_research/multimodal.py
@@ -1,0 +1,80 @@
+"""Multimodal content analysis for the Online Research agent.
+
+:class:`MultimodalAnalyzer` performs (currently mocked) image/video frame
+analysis plus text-based feature extraction shared across all source types
+(issue #1730).
+"""
+
+from __future__ import annotations
+
+import logging
+from typing import Dict, List
+
+logger = logging.getLogger(__name__)
+
+
+class MultimodalAnalyzer:
+    """Performs multimodal analysis on research content."""
+
+    def __init__(self):
+        # In production, would use actual LLM vision capabilities
+        pass
+
+    def analyze_image(self, image_url: str) -> Dict:
+        """Analyze an image for feature detection."""
+        print(f"Analyzing image: {image_url}")
+
+        # Mock analysis
+        return {
+            "detected_elements": ["block", "item", "UI element"],
+            "colors": ["#FF0000", "#00FF00", "#0000FF"],
+            "text_detected": False,
+            "confidence": 0.85,
+        }
+
+    def analyze_video_frame(self, frame_data: bytes) -> Dict:
+        """Analyze a video frame for features."""
+        # Mock analysis
+        return {
+            "detected_gameplay": "building",
+            "detected_items": ["stone", "wood"],
+            "detected_entities": ["player"],
+            "confidence": 0.78,
+        }
+
+    def extract_features_from_text(self, text: str) -> List[str]:
+        """Extract feature descriptions from text."""
+        features = []
+
+        # Common mod feature keywords
+        feature_keywords = [
+            "custom block",
+            "custom item",
+            "new dimension",
+            "new biome",
+            "crafting recipe",
+            "smelting",
+            "breeding",
+            "spawning",
+            "generation",
+            "structure",
+            "entity",
+            "tile entity",
+            "texture",
+            "model",
+            "sound",
+            "particle",
+            "effect",
+            "quest",
+            "achievement",
+            "skill",
+            "magic",
+            "tech",
+        ]
+
+        text_lower = text.lower()
+        for keyword in feature_keywords:
+            if keyword in text_lower:
+                features.append(keyword)
+
+        return list(set(features))
